Fix internal datatype usage and warnings (#573)
PR #568 increased the warning levels and c-ares code emitted a bunch of warnings. This PR fixes those warnings and starts transitioning internal data types into more proper forms (e.g. data lengths should be size_t not int). It does, however, have to manually cast back to what the public API needs due to API and ABI compliance (we aren't looking to break integrations, just clean up internals).

Fix By: Brad House (@bradh352)

+++ src/lib/ares__addrinfo_localhost.c
@@ -49,7 +49,7 @@
 
 ares_status_t ares_append_ai_node(int aftype,
                                   unsigned short port,
-                                  int ttl,
+                                  unsigned int ttl,
                                   const void *adata,
                                   struct ares_addrinfo_node **nodes)
 {
@@ -80,7 +80,7 @@ ares_status_t ares_append_ai_node(int aftype,
       node->ai_family = AF_INET;
       node->ai_addrlen = sizeof(*sin);
       node->ai_addr = (struct sockaddr *)sin;
-      node->ai_ttl = ttl;
+      node->ai_ttl = (int)ttl;
     }
 
   if (aftype == AF_INET6)
@@ -100,7 +100,7 @@ ares_status_t ares_append_ai_node(int aftype,
       node->ai_family = AF_INET6;
       node->ai_addrlen = sizeof(*sin6);
       node->ai_addr = (struct sockaddr *)sin6;
-      node->ai_ttl = ttl;
+      node->ai_ttl = (int)ttl;
     }
 
   return ARES_SUCCESS;
